JDWP: remove suspend count check on invoke
We used to return an error when the debugger asks to execute a method
in a thread that is suspended more than once. The reason was to avoid
blocking the JDWP thread on a thread that is still suspended.

Now invoke commands are handled asynchronously, we no longer need to
that check.

Bug: 19397712
Change-Id: I14f259923753e411dcce514183ed6fccd4cd0450
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>