enum operator<<() script fails with do { } while cond;
The script assumes that a line containing a '}' and a ';' will only
and always signal the end of a scope.
Unfortunately, this also matches '} while (myCond);',
thus incorrectly thinking it has reached the end of a scope.

This can result in enums being emitted without scope resolution
(or with insufficient scope resolution), cause compile errors.

Change-Id: I6e1b3453b941bff93ddd0b2a9fd3950182668728
Signed-off-by: randy.jeong <randy.jeong@samsung.com>
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>